--- packages/ui/react/src/components/Forms/Steppers/index.ts
@@ -1,1 +1,0 @@
-export * from './Steppers'
